Add terminate button (#54)
