+++ screens/profile/components.js
@@ -25,6 +25,7 @@ import {
 import {useTranslation} from 'react-i18next'
 import dayjs from 'dayjs'
 import {useMachine} from '@xstate/react'
+import {useQuery} from 'react-query'
 import {
   Avatar,
   Tooltip,
@@ -35,6 +36,7 @@ import {
   Input,
   DrawerFooter,
   Toast,
+  FormControlWithLabel,
 } from '../../shared/components/components'
 import {rem} from '../../shared/theme'
 import {PrimaryButton, SecondaryButton} from '../../shared/components/button'
@@ -66,6 +68,7 @@ import {
 import {useIdentity} from '../../shared/providers/identity-context'
 import {useEpoch} from '../../shared/providers/epoch-context'
 import {activateMiningMachine} from './machines'
+import {fetchBalance} from '../../shared/api/wallet'
 
 export function UserInlineCard({address, state}) {
   return (
@@ -692,6 +695,118 @@ export function DeactivateMiningDrawer({
             {t('Submit')}
           </PrimaryButton>
         </Stack>
+      </DrawerFooter>
+    </Drawer>
+  )
+}
+
+// eslint-disable-next-line react/prop-types
+export function KillForm({isOpen, onClose}) {
+  const {t} = useTranslation()
+  const {privateKey, coinbase} = useAuthState()
+  const [, {killMe}] = useIdentity()
+  const [submitting, setSubmitting] = useState(false)
+  const toast = useToast()
+
+  const [to, setTo] = useState()
+
+  const {
+    data: {stake},
+  } = useQuery(['get-balance', coinbase], () => fetchBalance(coinbase), {
+    initialData: {balance: 0, stake: 0},
+    enabled: !!coinbase,
+  })
+
+  const terminate = async () => {
+    try {
+      if (to !== coinbase)
+        throw new Error(t('You must specify your own identity address'))
+
+      setSubmitting(true)
+
+      await killMe(privateKey)
+
+      toast({
+        status: 'success',
+        // eslint-disable-next-line react/display-name
+        render: () => <Toast title={t('Transaction sent')} />,
+      })
+      if (onClose) onClose()
+    } catch (error) {
+      toast({
+        // eslint-disable-next-line react/display-name
+        render: () => (
+          <Toast
+            title={error?.message ?? t('error:Error while sending transaction')}
+            status="error"
+          />
+        ),
+      })
+    } finally {
+      setSubmitting(false)
+    }
+  }
+
+  return (
+    <Drawer isOpen={isOpen} onClose={onClose}>
+      <DrawerHeader mb={6}>
+        <Avatar address={coinbase} mx="auto" />
+        <Heading
+          fontSize="lg"
+          fontWeight={500}
+          color="brandGray.500"
+          mt={4}
+          mb={0}
+          textAlign="center"
+        >
+          {t('Terminate identity')}
+        </Heading>
+      </DrawerHeader>
+      <DrawerBody>
+        <Text fontSize="md" mb={6}>
+          {t(`Terminate your identity and withdraw the stake. Your identity status
+            will be reset to 'Not validated'.`)}
+        </Text>
+        <FormControlWithLabel label={t('Withraw stake, iDNA')}>
+          <Input value={stake} isDisabled />
+        </FormControlWithLabel>
+        <Text fontSize="md" mb={6} mt={6}>
+          {t(
+            'Please enter your identity address to confirm termination. Stake will be transferred to the identity address.'
+          )}
+        </Text>
+        <FormControlWithLabel label={t('Address')}>
+          <Input value={to} onChange={e => setTo(e.target.value)} />
+        </FormControlWithLabel>
+      </DrawerBody>
+      <DrawerFooter>
+        <Box
+          alignSelf="stretch"
+          borderTop="1px"
+          borderTopColor="gray.300"
+          mt="auto"
+          pt={5}
+          width="100%"
+        >
+          <Stack isInline spacing={2} justify="flex-end">
+            <PrimaryButton
+              onClick={terminate}
+              isLoading={submitting}
+              variantColor="red"
+              _hover={{
+                bg: 'rgb(227 60 60)',
+              }}
+              _active={{
+                bg: 'rgb(227 60 60)',
+              }}
+              _focus={{
+                boxShadow: '0 0 0 3px rgb(255 102 102 /0.50)',
+              }}
+            >
+              {t('Terminate')}
+            </PrimaryButton>
+          </Stack>
+        </Box>
       </DrawerFooter>
     </Drawer>
   )
